server/pyproject.toml
[project]
name = "qc-dashboard"
version = "0.1.0"
description = ""
authors = [
    {name = "Marcus Hennig",email = "marcus@planqc.eu"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn[standard] (>=0.34.2,<0.35.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "numpy (>=2.2.6,<3.0.0)"
]

[tool.poetry]
packages = [{include = "qc_dashboard", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
